Add skill string parsing function
This commit adds a function to parse a string containing a list of skills and return a list of skills. The function takes a string argument and uses multiple splitters to split the string into a list of skills. The function then determines the splitter that yields the most number of skills and returns that list. If no skills are found, an empty list is returned.